requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
pytest==5.1.2
pytest-cov==2.7.1
pytest-xdist==1.29.0
hypothesis==3.78.0
flaky==3.6.1
pytest-html==2.0.0
locustio==0.11.0
jsonschema==2.6.0
allure-pytest==2.8.5
requests
naughty-string-validator==0.1.2
mock==3.0.5
pytest-mock==1.12.1
